bindings/python: use Numpy 2.x; update wheel builds (#20)
This PR shifts the Python bindings from the Numpy 1.x to Numpy 2.x API.
See the following for context:

https://numpy.org/doc/stable/numpy_2_0_migration_guide.html
https://numpy.org/doc/2.0/dev/depending_on_numpy.html

Numpy 2.x is now a better choice than 1.x for modules:

- the ABI is backwards compatible (as far as 1.7), and
- newer Python versions are not compatible with 1.x.

It makes sense to shift getdata in the same direction too. Along the
way, I've refreshed the wheel build (to repair deprecated MacOS
builders, to enable Python 3.13, and to deprecate 3.8.)

Comments welcome. While the build succeeds and regression tests pass, I
have not used these code "in anger". I do plan to follow up this PR with
a wheel builder for Windows, and think Numpy 2.x is a better basis for
this work.